import/bikes: read mongo config from file
In a juju world, the mongodb instance will live on a random IP and
may have an arbitrary port.

The node-app charm writes this configuration into a file, and the nodejs
app already parses it properly.

If we want to do the same for our python importers, it's a little uglier,
but still possible.