charms/node-app: rejigger mongodb-relation-[joined|changed]
We only want to import data once, when the mongodb relation is joined.

However, all the configuration code for getting the address of the
mongodb server and its port lives in the 'relation-changed' hook, which
only runs *after* the 'relation-joined' hook.

We really want the node-app's config/config.js to be updated on both
'relation-joined' and 'relation-changed'. So in order to avoid repeating
ourselves, combine the scripts, add a bit of logic to detect which
hook we are called as, and then do the data import on the 'relation-joined'
hook only.