erver: lat/lng pairs should be query params, not resource locators
The point of OpenMotion is that users are *searching* for transport
modes near their current GPS coordinates. So it makes more semantic
sense for the API to accept lat/lng pairs as query parameters, and
not as resource locators (which are misleading).

A benefit to this approach is that we can also support additional
arguments in the future, such as search radius, etc.